Kinetis- Turn on USB buf to fixed dropped packets
When a SETUP packet comes immediately after a control transfer with a
data in stage, there is a chance that the SETUP packet will be
dropped. This is because no buffers are available to receive the setup
packet. This patch turns on double buffering to prevent SETUP packets
from being dropped.