Fix incorrect mass storage assert
In USBD_MSC_MemoryWrite if Offset and BulkLen are both 0 then then
there is an underflow and an assert is falsely triggered. This patch
corrects that logic error which prevents the false assert.

This occurs during a write if the media state changes to not ready
after the CBW has been read but before any data has been transferred.