stm32f103xb : Reduce chance of dropped setup pkts
When a SETUP packet comes immediately after a control transfer with a
data in stage, there is a chance that the SETUP packet will be
dropped. This is because no buffers are available to receive the setup
packet. This patch moves control ACK packet processing into the USB
ISR so the RX buffer is prepared for the next setup packet far faster.
This greatly reduces the chance of dropping a setup packet.