Implement and use manager for SWD
There are several components which access SWD independent of each
other. If two of these components use SWD at the same time then the
ongoing operation will fail and SWD can become corrupted.

To overcome this problem this patch adds an SWD manager which both
serializes access to SWD and adds an API for a component to request
access to SWD. The SWD manager code determines if SWD is already is
use and if it is it can decide to abort the current operation and
start the new operation, or prevent the new operation from starting.

The current implementation of this code always aborts the current
operation so the newly requested operation can start.